refactor `<MaintenanceBanner />` and associated code
