--- packages/manager/src/components/MaintenanceBanner/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MaintenanceBanner';
