+++ packages/manager/src/mocks/serverHandlers.ts
@@ -383,6 +383,7 @@ export const handlers = [
       })
   ),
   rest.get('*/linode/instances', async (req, res, ctx) => {
+    linodeFactory.resetSequenceNumber();
     const metadataLinodeWithCompatibleImage = linodeFactory.build({
       image: 'metadata-test-image',
       label: 'metadata-test-image',
@@ -450,6 +451,10 @@ export const handlers = [
       multipleIPLinode,
     ];
     return res(ctx.json(makeResourcePage(linodes)));
+  }),
+  rest.get('*/linode/instances/:id', async (req, res, ctx) => {
+    const id = Number(req.params.id);
+    return res(ctx.json(linodeFactory.build({ id })));
   }),
   rest.delete('*/instances/*', async (req, res, ctx) => {
     return res(ctx.json({}));
@@ -760,7 +765,7 @@ export const handlers = [
       headers.status === 'completed'
         ? accountMaintenanceFactory.buildList(30, { status: 'completed' })
         : [
-            ...accountMaintenanceFactory.buildList(2, { status: 'pending' }),
+            ...accountMaintenanceFactory.buildList(90, { status: 'pending' }),
             ...accountMaintenanceFactory.buildList(3, { status: 'started' }),
           ];
 
@@ -1242,6 +1247,7 @@ export const mockDataHandlers: Record<
 > = {
   linode: (count) =>
     rest.get('*/linode/instances', async (req, res, ctx) => {
+      linodeFactory.resetSequenceNumber();
       const linodes = linodeFactory.buildList(count);
       return res(ctx.json(makeResourcePage(linodes)));
     }),
