+++ packages/manager/src/utilities/isPast.ts
@@ -1,4 +1,4 @@
 import { parseAPIDate } from './date';
 
-export default (a: string) => (b: string): boolean =>
+export const isPast = (a: string) => (b: string): boolean =>
   parseAPIDate(b) >= parseAPIDate(a);
